% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareWithNationalStats.R
\name{compareDeliveryMode}
\alias{compareDeliveryMode}
\title{Compare delivery-mode distributions}
\usage{
compareDeliveryMode(deliveryModeSummary, natl)
}
\arguments{
\item{deliveryModeSummary}{App data (long format with mode, pct columns).}

\item{natl}{Parsed national statistics.}
}
\value{
A tibble.
}
\description{
Compare delivery-mode distributions
}
\keyword{internal}
