% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfToComment}
\alias{dfToComment}
\title{Convert a data.frame to an aligned, copy-pasteable R comment block}
\usage{
dfToComment(df)
}
\arguments{
\item{df}{A data.frame / tibble}
}
\value{
Invisibly returns df (printed as comments)
}
\description{
Convert a data.frame to an aligned, copy-pasteable R comment block
}
