% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{export}
\usage{
export(cdm, outputDir, exportDir, minCellCount = 5)
}
\arguments{
\item{cdm}{(`cdm_reference`) CDM-Reference}

\item{outputDir}{(`character(1)`) The path to the outputs generated by `runHipps()`}

\item{exportDir}{(`character(1)`) The path to the folder where the sharable outputs should be written to.}

\item{minCellCount}{minimum counts}
}
\value{
`NULL`
}
\description{
Exports the patient level results to summarised share-able csv-files.
}
