% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertPregnancyEpisodesTable.R
\name{insertPregnancyEpisodesTable}
\alias{insertPregnancyEpisodesTable}
\title{insertPregnancyEpisodesTable}
\usage{
insertPregnancyEpisodesTable(cdm, path, tableName, overwrite = FALSE)
}
\arguments{
\item{cdm}{(`cdm_reference`)}

\item{path}{(`character(1)`) Path to the generated rds-file generated by `runHipps()`.}

\item{tableName}{(`character(1)`) Name of the table}

\item{overwrite}{(`logical(1)`: `TRUE`) If the table should be overwritten or not.}
}
\value{
(`cdm_reference`)
}
\description{
Inserts the identified pregnancy episodes table into a database.
}
