% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPregnancyIdentifier.R
\name{makeLogger}
\alias{makeLogger}
\title{Create a new logger}
\usage{
makeLogger(outputDir, outputLogToConsole = TRUE)
}
\arguments{
\item{outputDir}{The directory where the log should be created}

\item{outputLogToConsole}{(`logical(1)`) If `TRUE` (default), log messages are
written to the console as well as the log file. If `FALSE`, only the file
appender is used (useful in tests to keep output clean).}
}
\value{
A log4r logger object
}
\description{
Create a new logger
}
