% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeHIPPSEpisodes.R
\name{mergeHipPps}
\alias{mergeHipPps}
\title{mergeHipPps}
\usage{
mergeHipPps(HIP, PPSEpisode, PPSMinMax, cdm, outputDir, ...)
}
\arguments{
\item{HIP}{(`data.frame`) HIP result}

\item{PPSEpisode}{(`data.frame`)}

\item{PPSMinMax}{(`data.frame`)}

\item{cdm}{(`cdm_reference`)}

\item{outputDir}{(`character(1)`)}

\item{...}{Extra (dev) parameters}
}
\value{
`NULL`
}
\description{
mergeHipPps
}
