% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mockPregnancyCdm}
\alias{mockPregnancyCdm}
\title{Create a mock pregnancy cdm for examples and testing}
\usage{
mockPregnancyCdm()
}
\value{
A cdm reference with some example pregnancy data in it
}
\description{
Create a mock pregnancy cdm for examples and testing
}
\examples{
\dontrun{
cdm <- mockPregnancyCdm()
}
}
