% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareWithNationalStats.R
\name{parseEuropeanNumber}
\alias{parseEuropeanNumber}
\title{Parse a character vector of numbers that may use European formatting
(dots as thousands separators, commas as decimal separators).}
\usage{
parseEuropeanNumber(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Numeric vector.
}
\description{
Parse a character vector of numbers that may use European formatting
(dots as thousands separators, commas as decimal separators).
}
\keyword{internal}
