% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PregnancyIdentifier-package.R
\docType{package}
\name{PregnancyIdentifier-package}
\alias{PregnancyIdentifier}
\alias{PregnancyIdentifier-package}
\title{PregnancyIdentifier: Identifies Pregnancy Episodes in the OMOP (Observational Medical Outcomes Partnership) CDM (Common Data Model)}
\description{
Identify pregnancy episodes in the OMOP CDM, based on the HIPPS algorithm described in Smith et al. (2024) \doi{10.1093/jamia/ocae195}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://darwin-eu.github.io/PregnancyIdentifier/}
  \item Report bugs at \url{https://darwin-eu.github.io/PregnancyIdentifier/issues/}
}

}
\author{
\strong{Maintainer}: Maarten van Kessel \email{m.l.vankessel@erasmusmc.com} (\href{https://orcid.org/0009-0006-8832-6030}{ORCID})

Authors:
\itemize{
  \item Aniek Markus (\href{https://orcid.org/0000-0001-5779-4794}{ORCID})
  \item Julieta Politi (\href{https://orcid.org/0000-0001-7315-034X}{ORCID})
  \item Adam Black \email{black@ohdsi.org} (\href{https://orcid.org/0000-0001-5576-8701}{ORCID})
}

}
\keyword{internal}
