% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESD.R
\name{runEsd}
\alias{runEsd}
\title{runEsd}
\usage{
runEsd(
  HIPPS,
  cdm,
  outputDir,
  uploadConceptSets = FALSE,
  startDate = as.Date("1900-01-01"),
  endDate = Sys.Date(),
  logger,
  ...
)
}
\arguments{
\item{HIPPS}{(`data.frame`)}

\item{cdm}{(`cdm_reference`)}

\item{outputDir}{(`character(1)`)}

\item{uploadConceptSets}{if concept sets should be uploaded}

\item{startDate}{(`Date(1)`: `as.Date("1900-01-01"`) Start date of data to use. By default 1900-01-01}

\item{endDate}{(`Date(1)`: `Sys.Date()`) End date of data to use. By default today.}

\item{logger}{(`logger`) Logger object.}

\item{...}{Extra (development) parameters}
}
\value{
`NULL`
}
\description{
Runs ESD
}
