% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP.R
\name{runHip}
\alias{runHip}
\title{runHip}
\usage{
runHip(
  cdm,
  outputDir,
  startDate = as.Date("1900-01-01"),
  endDate = Sys.Date(),
  logger,
  ...
)
}
\arguments{
\item{cdm}{(`cdm_reference`) A CDM-Reference object from CDMConnector.}

\item{outputDir}{(`character(1)`) Output directory to write output to.}

\item{startDate}{(`Date(1)`: `as.Date("1900-01-01"`) Start date of data to use. By default 1900-01-01}

\item{endDate}{(`Date(1)`: `Sys.Date()`) End date of data to use. By default today.}

\item{logger}{(`logger`) Logger object.}

\item{...}{Dev params}
}
\value{
cdm object
}
\description{
Runs the HIP
}
