% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initPregnancies.R
\name{initPregnancies}
\alias{initPregnancies}
\title{Create Initial Pregnancy Table}
\usage{
initPregnancies(
  cdm,
  startDate = as.Date("1900-01-01"),
  endDate = Sys.Date(),
  ageBounds = c(15L, 56L)
)
}
\arguments{
\item{cdm}{A CDM reference}

\item{startDate}{Earliest date to look for pregnancies in the CDM}

\item{endDate}{Latest date to look for pregnancies in the CDM}

\item{ageBound}{The upper and lower bounds for age at pregnancy end date
represented using a length 2 integer vector. By default this will be
c(15, 56) and will include anyone >= 15 and < 56.}
}
\value{
The input CDM with a new table added called preg_initial_cohort with
the following columns:"person_id" "concept_id" "visit_date" "value_as_number" "concept_name"
"category" "gest_value" "date_of_birth" "date_diff" "age"
}
\description{
Get initial cohort based on HIP concepts for women who are age 15 to 55
at the time of pregnancy
}
