% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initPregnancies.R
\name{initPregnancies}
\alias{initPregnancies}
\title{Create Initial Pregnancy Table}
\usage{
initPregnancies(
  cdm,
  startDate = as.Date("1900-01-01"),
  endDate = Sys.Date(),
  ageBounds = c(15L, 56L),
  logger,
  outputDir = NULL
)
}
\arguments{
\item{cdm}{A CDM reference}

\item{startDate}{Earliest date to look for pregnancies in the CDM}

\item{endDate}{Latest date to look for pregnancies in the CDM}

\item{ageBounds}{The upper and lower bounds for age at pregnancy end date
represented using a length 2 integer vector. By default this will be
c(15, 56) and will include anyone >= 15 and < 56.}

\item{logger}{A log4r logger object that can be created with `makeLogger()`}

\item{outputDir}{Optional directory path. When provided, an \code{attrition.csv}
file is created with initial record and person counts for \code{preg_hip_records}
and \code{preg_pps_records}.}
}
\value{
The input CDM with new tables added called preg_hip_records, preg_pps_records,
preg_hip_concepts, preg_pps_concepts.
}
\description{
Get initial cohort based on HIP concepts for women who are age 15 to 55
at the time of pregnancy
}
\examples{
\dontrun{
logger <- makeLogger(tempdir())
cdm <- mockPregnancyCdm()
cdm <- initPregnancies(cdm, logger = logger)
}
}
