% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPregnancyIdentifier.R
\name{makeLogger}
\alias{makeLogger}
\title{Create a new logger}
\usage{
makeLogger(outputDir)
}
\arguments{
\item{outputDir}{The directory where the log should be created}
}
\value{
A log4r logger object
}
\description{
Create a new logger
}
