% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESD.R
\name{runEsd}
\alias{runEsd}
\title{runEsd}
\usage{
runEsd(HIPPS, cdm, outputDir, uploadConceptSets = FALSE, logger, ...)
}
\arguments{
\item{HIPPS}{(`data.frame`)}

\item{cdm}{(`cdm_reference`)}

\item{outputDir}{(`character(1)`)}

\item{uploadConceptSets}{if concept sets should be uploaded}

\item{...}{Extra (development) parameters}
}
\value{
`NULL`
}
\description{
Runs ESD
}
