% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHIP.R
\name{runHip}
\alias{runHip}
\title{Run the HIP pregnancy identification algorithm}
\usage{
runHip(
  cdm,
  outputDir = NULL,
  startDate = as.Date("1900-01-01"),
  endDate = Sys.Date(),
  justGestation = TRUE,
  logger = NULL
)
}
\arguments{
\item{cdm}{(`cdm_reference`) A cdm reference object from CDMConnector.}

\item{outputDir}{(`character(1)`) Output directory to write output to.}

\item{startDate}{(`Date(1)`: `as.Date("1900-01-01"`) Start date of data to use. By default 1900-01-01}

\item{endDate}{(`Date(1)`: `Sys.Date()`) End date of data to use. By default today.}

\item{justGestation}{(`logical(1)`: `TRUE`) Should episodes that only have gestational concepts be considered?}

\item{logger}{(`logger`) A log4r logger object.}
}
\value{
The input cdm object with cdm$preg_hip_episodes table added.
Columns of this table are:
- person_id: Unique person identifier
- hip_episode: number of the episode within person_id (1=first, 2=second, etc.)
- hip_pregnancy_start: Estimated start date of the pregnancy episode
- hip_pregnancy_end: Date of the pregnancy outcome (i.e. pregnancy end date)
- hip_outcome_category: HIP pregnancy outcome category
- hip_first_gest_date: First gestation date in the episode (NA if not identified)
- hip_gest_flag: Were gestational age concepts found in the episode? yes or no
- hip_episode_length: length of the episode in days
}
\description{
This function runs the HIP outcome based pregnancy identification algorithm.
It assumes that the `initPregnancies` function has already been run and the
preg_hip_records and preg_hip_concepts tables are in the cdm. The algorithm
identifies start and end dates for pregnancies by first prioritizing pregnancy
outcomes and then using inference to set a pregnancy start date. Each pregnancy
is classified into one of several categories.
- "LB": Live birth
- "SB": Still birth
- "AB": abortion,
- "SA": miscarriage,
- "DELIV": unspecified delivery
- "ECT": ectopic pregnancy
- "PREG": unspecified/ongoing pregnancy
}
