% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP.R
\name{runHip}
\alias{runHip}
\title{runHip}
\usage{
runHip(cdm, outputDir, logger, ...)
}
\arguments{
\item{cdm}{(`cdm_reference`) A CDM-Reference object from CDMConnector.}

\item{outputDir}{(`character(1)`) Output directory to write output to.}

\item{...}{Dev params}
}
\value{
cdm object
}
\description{
Runs the HIP
}
