% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHIPPS.R
\name{runHipps}
\alias{runHipps}
\title{runHipps}
\usage{
runHipps(cdm, outputDir, fileName)
}
\arguments{
\item{cdm}{(`cdm_reference`) A CDM-Reference object from CDMConnector.}

\item{outputDir}{(`character(1)`) Output directory to write output to.}

\item{fileName}{(`character(1)`) Filename to write to}
}
\value{
`NULL`
}
\description{
Runs the HIPPS algorithm (HIP, PPS, and ESD) from: https://github.com/louisahsmith/allofus-pregnancy/
without requiring the `allofus` R package.
}
