% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHIPPS.R
\name{runHipps}
\alias{runHipps}
\title{runHipps}
\usage{
runHipps(
  cdm,
  outputDir,
  startDate = as.Date("1900-01-01"),
  endDate = Sys.Date()
  justGestation = TRUE,
  ...
)
}
\arguments{
\item{cdm}{(`cdm_reference`) A CDM-Reference object from CDMConnector.}

\item{outputDir}{(`character(1)`) Output directory to write output to.}

\item{startDate}{(`Date(1)`: `as.Date("1900-01-01"`) Start date of data to use. By default 1900-01-01}

\item{endDate}{(`Date(1)`: `Sys.Date()`) End date of data to use. By default today.}

\item{justGestation}{(`logical(1)`: `TRUE`) Should episodes that only have gestational concepts be concidered?}

\item{...}{Dev params}
}
\value{
`NULL`
}
\description{
Runs the HIPPS algorithm (HIP, PPS, and ESD) from: https://github.com/louisahsmith/allofus-pregnancy/
without requiring the `allofus` R package.
}
