% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPS.R
\name{runPps}
\alias{runPps}
\title{runPps}
\usage{
runPps(cdm, outputDir, uploadConceptSets = FALSE, logger, ...)
}
\arguments{
\item{cdm}{(`cdm_reference`)}

\item{outputDir}{output directory}

\item{uploadConceptSets}{if concept sets should be uploaded}

\item{logger}{(`logger`) Logger object.}

\item{...}{optional parameters}
}
\value{
cdm object
}
\description{
Runs PPS algorithm
}
